Pass CLEANUP=0 to test_e2e to leave logs

+++ test_e2e.sh
@@ -2109,5 +2109,10 @@ done
 
 # Finally...
 echo
-echo "all tests passed: cleaning up $DIR"
-rm -rf "$DIR"
+if [[ "${CLEANUP:-}" == 0 ]]; then
+    echo "leaving logs in $DIR"
+else
+    echo "cleaning up $DIR"
+    rm -rf "$DIR"
+fi
+
